+++ src/nn-builder/src/data_structure/ndarray.h
@@ -11,17 +11,17 @@ namespace ds {
 class NDArray {
 private:
   wasmpp::Memory* memory_;
-  std::vector<uint64_t> shape_;
-  uint64_t unit_size_;
+  std::vector<uint32_t> shape_;
+  uint32_t unit_size_;
   // Store the multiplication value for the shape elements
   // left to right. This is useful for computing the index
   // in the linear memory
-  std::vector<uint64_t> shape_mul_;
+  std::vector<uint32_t> shape_mul_;
 public:
-  NDArray(wasmpp::Memory* memory, std::vector<uint64_t> shape, uint64_t unit_size);
-  void Reshape(std::vector<uint64_t> shape);
-  std::vector<uint64_t >Shape() const { return shape_;}
-  uint64_t GetLinearIndex(std::vector<uint64_t> index) const;
+  NDArray(wasmpp::Memory* memory, std::vector<uint32_t> shape, uint32_t unit_size);
+  void Reshape(std::vector<uint32_t> shape);
+  std::vector<uint32_t >Shape() const { return shape_;}
+  uint32_t GetLinearIndex(std::vector<uint32_t> index) const;
   const wasmpp::Memory* Memory() const { return memory_; }
 };
